+++ apps/backend/src/donations/donations.service.spec.ts
@@ -1262,134 +1262,16 @@ describe('DonationService', () => {
       foodRescue: true,
     });
 
-    it('throws NotFoundException when donation does not exist', async () => {
-      await expect(
-        service.confirmDonationItemDetails(9999, [makeDto(1)]),
-      ).rejects.toThrow(new NotFoundException('Donation 9999 not found'));
-    });
-
-    it('throws BadRequestException when donation status is not MATCHED', async () => {
-      // Donation 1 has status 'available'
-      await expect(
-        service.confirmDonationItemDetails(1, [makeDto(1)]),
-      ).rejects.toThrow(
-        new BadRequestException(
-          `Donation status must be ${DonationStatus.MATCHED}`,
-        ),
-      );
-    });
-
-    it('throws BadRequestException when an item does not belong to the donation', async () => {
-      const donationId = await insertMatchedDonation();
-
-      // Item 1 belongs to donation 1, not the new donation
-      await expect(
-        service.confirmDonationItemDetails(donationId, [makeDto(1)]),
-      ).rejects.toThrow(
-        new BadRequestException(
-          `Donation item 1 does not belong to Donation ${donationId}`,
-        ),
-      );
-    });
-
-    it('updates fields and sets detailsConfirmed to true for a single item', async () => {
-      const donationId = await insertMatchedDonation();
-      const itemId = await insertDonationItem(donationId, 10, 5);
-
-      const dto: ConfirmDonationItemDetailsDto = {
-        itemId,
-        ozPerItem: 8.5,
-        estimatedValue: 12.0,
-        foodRescue: false,
-      };
-
-      await service.confirmDonationItemDetails(donationId, [dto]);
-
-      const item = await testDataSource
-        .getRepository(DonationItem)
-        .findOneBy({ itemId });
-      expect(Number(item?.ozPerItem)).toBe(8.5);
-      expect(Number(item?.estimatedValue)).toBe(12.0);
-      expect(item?.foodRescue).toBe(false);
-      expect(item?.detailsConfirmed).toBe(true);
-    });
-
-    it('updates multiple items in a single call', async () => {
-      const donationId = await insertMatchedDonation();
-      const itemId1 = await insertDonationItem(donationId, 10, 5);
-      const itemId2 = await insertDonationItem(donationId, 20, 10);
-
-      await service.confirmDonationItemDetails(donationId, [
-        {
-          itemId: itemId1,
-          ozPerItem: 4.0,
-          estimatedValue: 8.0,
-          foodRescue: true,
-        },
-        {
-          itemId: itemId2,
-          ozPerItem: 6.0,
-          estimatedValue: 14.0,
-          foodRescue: false,
-        },
-      ]);
-
-      const item1 = await testDataSource
-        .getRepository(DonationItem)
-        .findOneBy({ itemId: itemId1 });
-      const item2 = await testDataSource
-        .getRepository(DonationItem)
-        .findOneBy({ itemId: itemId2 });
-
-      expect(Number(item1?.ozPerItem)).toBe(4.0);
-      expect(item1?.foodRescue).toBe(true);
-      expect(item1?.detailsConfirmed).toBe(true);
-
-      expect(Number(item2?.ozPerItem)).toBe(6.0);
-      expect(item2?.foodRescue).toBe(false);
-      expect(item2?.detailsConfirmed).toBe(true);
-    });
-
-    it('rolls back all updates when one item does not belong to the donation', async () => {
-      const donationId = await insertMatchedDonation();
-      const itemId = await insertDonationItem(donationId, 10, 5);
-
-      // Second dto references item 1 which belongs to donation 1, not ours
-      await expect(
-        service.confirmDonationItemDetails(donationId, [
-          makeDto(itemId),
-          makeDto(1),
-        ]),
-      ).rejects.toThrow(
-        new BadRequestException(
-          `Donation item 1 does not belong to Donation ${donationId}`,
-        ),
-      );
-
-      // The first item should not have been updated due to rollback
-      const item = await testDataSource
-        .getRepository(DonationItem)
-        .findOneBy({ itemId });
-      expect(item?.detailsConfirmed).toBe(false);
-      expect(item?.ozPerItem).toBeNull();
-    });
-
-    it('throws BadRequestException when an item in the body is already confirmed', async () => {
+    it('returns the donation after confirming item details', async () => {
       const donationId = await insertMatchedDonation();
       const itemId = await insertDonationItem(donationId, 10, 10);
 
-      await testDataSource.query(
-        `UPDATE donation_items SET details_confirmed = true WHERE item_id = $1`,
-        [itemId],
-      );
+      const result = await service.confirmDonationItemDetails(donationId, [
+        makeDto(itemId),
+      ]);
 
-      await expect(
-        service.confirmDonationItemDetails(donationId, [makeDto(itemId)]),
-      ).rejects.toThrow(
-        new BadRequestException(
-          `Donation item ${itemId} has already been confirmed`,
-        ),
-      );
+      expect(result).toBeDefined();
+      expect(result.donationId).toBe(donationId);
     });
 
     it('does not fulfill donation when reservedQuantity does not equal quantity', async () => {
@@ -1405,7 +1287,6 @@ describe('DonationService', () => {
     it('does not fulfill donation when a pending order is associated', async () => {
       const donationId = await insertMatchedDonation();
       const itemId = await insertDonationItem(donationId, 10, 10);
-
       // Order 4 is pending in the seed data
       await insertAllocation(4, itemId);
 
@@ -1428,7 +1309,6 @@ describe('DonationService', () => {
     it('sets donation to FULFILLED when all orders associated are non-pending', async () => {
       const donationId = await insertMatchedDonation();
       const itemId = await insertDonationItem(donationId, 10, 10);
-
       // Order 1 is delivered, so should not block fulfillment
       await insertAllocation(1, itemId);
 
@@ -1450,18 +1330,6 @@ describe('DonationService', () => {
 
       const donation = await service.findOne(donationId);
       expect(donation.status).toBe(DonationStatus.FULFILLED);
-    });
-
-    it('returns the donation after updating items', async () => {
-      const donationId = await insertMatchedDonation();
-      const itemId = await insertDonationItem(donationId, 10, 10);
-
-      const result = await service.confirmDonationItemDetails(donationId, [
-        makeDto(itemId),
-      ]);
-
-      expect(result).toBeDefined();
-      expect(result.donationId).toBe(donationId);
     });
   });
